fix: correct architectures
The architectures in the charmcraft.yaml aren't correct. We've got
duplicate arm64 architectures (arm64 and aarch64), so we should
just pick one. Power PC was completely missing.

This charm currently only works by happenstance because what we
build on is only amd64. If there were any architectual specific
binaries we would stop working as we don't correctly build for those.