feat: write controller config atomically
It's possible to corrupt the controller config file if there is an
exception during writing of the file. Instead, we want to only update
the controller config atomically. This involves writing the yaml to
a temp file and then do a atomic file replace (POSIX). This ensures
that it will always be valid if the data is valid.